fix: dataset publisher
