+++ apis/dataset_metadata/dataset_other.py
@@ -80,56 +80,3 @@ class DatasetOtherResource(Resource):
         dataset_.dataset_other.update(data)
         model.db.session.commit()
         return dataset_.dataset_other.to_dict(), 200
-
-
-@api.route("/study/<study_id>/dataset/<dataset_id>/metadata/publisher")
-class DatasetPublisherResource(Resource):
-    """Dataset Publisher Resource"""
-
-    @api.doc("publisher")
-    @api.response(200, "Success")
-    @api.response(400, "Validation Error")
-    # @api.marshal_with(dataset_publisher)
-    def get(self, study_id: int, dataset_id: int):  # pylint: disable= unused-argument
-        """Get dataset publisher metadata"""
-        dataset_ = model.Dataset.query.get(dataset_id)
-        dataset_other_ = dataset_.dataset_other
-        return dataset_other_.to_dict(), 200
-
-    @api.doc("update publisher")
-    @api.response(200, "Success")
-    @api.response(400, "Validation Error")
-    def put(self, study_id: int, dataset_id: int):
-        """Update dataset publisher metadata"""
-        study_obj = model.Study.query.get(study_id)
-
-        if not is_granted("dataset_metadata", study_obj):
-            return "Access denied, you can not make any change in dataset metadata", 403
-
-        schema = {
-            "type": "object",
-            "additionalProperties": False,
-            "properties": {
-                "publisher": {"type": "string", "minLength": 1},
-                "managing_organization_name": {"type": "string", "minLength": 1},
-                "managing_organization_ror_id": {
-                    "type": "string",
-                },
-            },
-            "required": [
-                "publisher",
-                "managing_organization_name",
-                "managing_organization_ror_id",
-            ],
-        }
-
-        try:
-            validate(instance=request.json, schema=schema)
-        except ValidationError as err:
-            return err.message, 400
-
-        data = request.json
-        dataset_ = model.Dataset.query.get(dataset_id)
-        dataset_.dataset_other.update(data)
-        model.db.session.commit()
-        return dataset_.dataset_other.to_dict(), 200
